requirements.txt
edge-tts
asyncio
pygame
openai-whisper
sounddevice
scipy
numpy
requests
psutil
pyperclip
feedparser
setuptools
